+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 0.7.2 (in progress)
 * Date fields now work with ISO 8601 strings rather than unix timestamps
+* Fix bugs with lists of microschemas (SUP-4712)
 
 ## Version 0.7.1 (2017-07-07)
 * Synchronize version with maven